+++ IntuneBackupAndRestore/Public/Invoke-IntuneBackupDeviceCompliancePolicyAssignment.ps1
@@ -16,8 +16,18 @@ function Invoke-IntuneBackupDeviceCompliancePolicyAssignment {
     [CmdletBinding()]
     param(
         [Parameter(Mandatory = $true)]
-        [string]$Path
+        [string]$Path,
+
+        [Parameter(Mandatory = $false)]
+        [ValidateSet("v1.0", "Beta")]
+        [string]$ApiVersion = "Beta"
     )
+
+    # Set the Microsoft Graph API endpoint
+    if (-not ((Get-MSGraphEnvironment).SchemaVersion -eq $apiVersion)) {
+        Update-MSGraphEnvironment -SchemaVersion $apiVersion -Quiet
+        Connect-MSGraph -ForceNonInteractive -Quiet
+    }
 
     # Create folder if not exists
     if (-not (Test-Path "$Path\Device Compliance Policies\Assignments")) {
@@ -25,10 +35,10 @@ function Invoke-IntuneBackupDeviceCompliancePolicyAssignment {
     }
 
     # Get all assignments from all policies
-    $deviceCompliancePolicies = Get-GraphDeviceCompliancePolicy
+    $deviceCompliancePolicies = Get-DeviceManagement_DeviceCompliancePolicies | Get-MSGraphAllPages
 
     foreach ($deviceCompliancePolicy in $deviceCompliancePolicies) {
-        $assignments = Get-GraphDeviceCompliancePolicyAssignment -Id $deviceCompliancePolicy.id 
+        $assignments = Get-DeviceManagement_DeviceCompliancePolicies_Assignments -DeviceCompliancePolicyId $deviceCompliancePolicy.id 
         if ($assignments) {
             Write-Output "Backing Up - Device Compliance Policy - Assignments: $($deviceCompliancePolicy.displayName)"
             $fileName = ($deviceCompliancePolicy.displayName).Split([IO.Path]::GetInvalidFileNameChars()) -join '_'